updated crud_app.py
had some issues with the db not displaying relationships and storing old nodes interfering with new builds so made it so the db will clear old nodes to interfere with script when running new.  also added checks to make sure things are successful so the logs will tell if anything is not working from script.